+++ deploy/images.env
@@ -29,9 +29,9 @@ CSI_SIG_STORAGE_PROVISIONER_IMAGE=registry.k8s.io/sig-storage/csi-provisioner@sh
 KUBE_RBAC_PROXY_IMAGE=quay.io/openshift/origin-kube-rbac-proxy@sha256:e2def4213ec0657e72eb790ae8a115511d5b8f164a62d3568d2f1bff189917e8
 KUBEVIRT_CONSOLE_PLUGIN_IMAGE=quay.io/kubevirt-ui/kubevirt-plugin@sha256:4ed1f37116cbaf883edd8783201d091fbbe121fcb84e7184ee531d3b1abd77ab
 KUBEVIRT_CONSOLE_PROXY_IMAGE=quay.io/kubevirt-ui/kubevirt-apiserver-proxy@sha256:38293a1050bce9c7cb786cbe8719be1edb91a922a6c583eedb61e56729ed4c72
-AAQ_OPERATOR_IMAGE=quay.io/kubevirt/aaq-operator@sha256:697b5292877f92c5af0e3cb576c31d80446610939bf550bde323e08ca0994e42
-AAQ_SERVER_IMAGE=quay.io/kubevirt/aaq-server@sha256:4428871e2133b33de0f2c9a6c904b6de45bccbd64f53505268bf76e64a60ab74
-AAQ_CONTROLLER_IMAGE=quay.io/kubevirt/aaq-controller@sha256:99dc6d1f6dc1fa03fc6caf32a10dd6e295446c1c2672de61cc4d3f78e86a3c29
+AAQ_OPERATOR_IMAGE=quay.io/kubevirt/aaq-operator@sha256:2d1c49702fd187030dfa669791f25cefbc29e3103c2f5e0318af4c0ccead0fd2
+AAQ_SERVER_IMAGE=quay.io/kubevirt/aaq-server@sha256:827356ed0ad3857c1319b30cd026775dc8076eb9a1b2b62338f617429af36381
+AAQ_CONTROLLER_IMAGE=quay.io/kubevirt/aaq-controller@sha256:088da1c52a1b14edff7e9d314a46bd9c09c780b39ff82038fb90dc1022ae630d
 DIGEST_LIST="quay.io/kubevirt/virt-operator@sha256:0e6d56ea0ebdf9bb832f29ab0dc4603f0f619474bb8bc02a01e0a06e9d8d6f4a"
 DIGEST_LIST="${DIGEST_LIST},quay.io/kubevirt/virt-api@sha256:735662c824da3d6069624d40674163d758c2256f8043cf6f02f505c3a724afa3"
 DIGEST_LIST="${DIGEST_LIST},quay.io/kubevirt/virt-controller@sha256:4f576e30b83db7b9f96dde77896898bde87db3517963c4227634d0c3e96bb86d"
@@ -63,6 +63,6 @@ DIGEST_LIST="${DIGEST_LIST},registry.k8s.io/sig-storage/csi-provisioner@sha256:d
 DIGEST_LIST="${DIGEST_LIST},quay.io/openshift/origin-kube-rbac-proxy@sha256:e2def4213ec0657e72eb790ae8a115511d5b8f164a62d3568d2f1bff189917e8"
 DIGEST_LIST="${DIGEST_LIST},quay.io/kubevirt-ui/kubevirt-plugin@sha256:4ed1f37116cbaf883edd8783201d091fbbe121fcb84e7184ee531d3b1abd77ab"
 DIGEST_LIST="${DIGEST_LIST},quay.io/kubevirt-ui/kubevirt-apiserver-proxy@sha256:38293a1050bce9c7cb786cbe8719be1edb91a922a6c583eedb61e56729ed4c72"
-DIGEST_LIST="${DIGEST_LIST},quay.io/kubevirt/aaq-operator@sha256:697b5292877f92c5af0e3cb576c31d80446610939bf550bde323e08ca0994e42"
-DIGEST_LIST="${DIGEST_LIST},quay.io/kubevirt/aaq-server@sha256:4428871e2133b33de0f2c9a6c904b6de45bccbd64f53505268bf76e64a60ab74"
-DIGEST_LIST="${DIGEST_LIST},quay.io/kubevirt/aaq-controller@sha256:99dc6d1f6dc1fa03fc6caf32a10dd6e295446c1c2672de61cc4d3f78e86a3c29"
+DIGEST_LIST="${DIGEST_LIST},quay.io/kubevirt/aaq-operator@sha256:2d1c49702fd187030dfa669791f25cefbc29e3103c2f5e0318af4c0ccead0fd2"
+DIGEST_LIST="${DIGEST_LIST},quay.io/kubevirt/aaq-server@sha256:827356ed0ad3857c1319b30cd026775dc8076eb9a1b2b62338f617429af36381"
+DIGEST_LIST="${DIGEST_LIST},quay.io/kubevirt/aaq-controller@sha256:088da1c52a1b14edff7e9d314a46bd9c09c780b39ff82038fb90dc1022ae630d"
