+++ hack/config
@@ -8,7 +8,7 @@ TTO_VERSION="v0.5.0"
 HPPO_VERSION="v0.18.0"
 HPP_VERSION="v0.18.0"
 MTQ_VERSION="v1.2.0"
-AAQ_VERSION="v1.1.2"
+AAQ_VERSION="v1.1.4"
 
 NODE_DRIVER_REG_VERSION="v2.10.0"
 LIVENESS_PROBE_VERSION="v2.12.0"
