+++ tests/go.mod
@@ -41,7 +41,7 @@ require (
 	github.com/openshift/custom-resource-status v1.1.2
 	gopkg.in/yaml.v3 v3.0.1
 	k8s.io/apiserver v0.28.4
-	kubevirt.io/applications-aware-quota v1.1.2
+	kubevirt.io/application-aware-quota v1.1.4
 	kubevirt.io/controller-lifecycle-operator-sdk/api v0.2.4
 	kubevirt.io/managed-tenant-quota v1.2.0
 	kubevirt.io/ssp-operator/api v0.19.0
